+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Reputation command
+- Leaderboard command
 
 ## [1.0.0] - 2019-04-03
 ### Added